+++ desc/objectives/_omnigenity.py
@@ -2,7 +2,8 @@
 
 import warnings
 
-from desc.backend import jnp, vmap
+from desc.backend import jnp
+from desc.batching import vmap_chunked
 from desc.compute import get_profiles, get_transforms
 from desc.compute._omnigenity import _omnigenity_mapping
 from desc.compute.utils import _compute as compute_fun
@@ -31,6 +32,10 @@ class QuasisymmetryBoozer(_Objective):
         Poloidal resolution of Boozer transformation. Default = 2 * eq.M.
     N_booz : int, optional
         Toroidal resolution of Boozer transformation. Default = 2 * eq.N.
+    surf_batch_size: int
+        Number of flux surfaces to compute simultaneously. Defaults to
+        computing all flux surfaces simultaneously. Decrease to reduce
+        memory required for computation.
 
     """
 
@@ -40,7 +45,7 @@ class QuasisymmetryBoozer(_Objective):
 
     _units = "(T)"
     _print_value_fmt = "Quasi-symmetry Boozer error: "
-    _static_attrs = _Objective._static_attrs + ["_helicity"]
+    _static_attrs = _Objective._static_attrs + ["_helicity", "_surf_batch_size"]
 
     def __init__(
         self,
@@ -58,6 +63,7 @@ class QuasisymmetryBoozer(_Objective):
         N_booz=None,
         name="QS Boozer",
         jac_chunk_size=None,
+        surf_batch_size=None,
     ):
         if target is None and bounds is None:
             target = 0
@@ -65,6 +71,7 @@ class QuasisymmetryBoozer(_Objective):
         self.helicity = helicity
         self.M_booz = M_booz
         self.N_booz = N_booz
+        self._surf_batch_size = surf_batch_size
         super().__init__(
             things=eq,
             target=target,
@@ -142,6 +149,7 @@ class QuasisymmetryBoozer(_Objective):
             "profiles": profiles,
             "matrix": matrix,
             "idx": idx,
+            "surf_batch_size": self._surf_batch_size,
         }
 
         timer.stop("Precomputing transforms")
@@ -180,6 +188,7 @@ class QuasisymmetryBoozer(_Objective):
             params=params,
             transforms=constants["transforms"],
             profiles=constants["profiles"],
+            surf_batch_size=constants["surf_batch_size"],
         )
         B_mn = data["|B|_mn_B"].reshape((constants["transforms"]["grid"].num_rho, -1))
         B_mn = constants["matrix"] @ B_mn.T
@@ -538,6 +547,10 @@ class Omnigenity(_Objective):
         computation time during optimization and only ``eq`` is allowed to change.
         If False, the field is allowed to change during the optimization and its
         associated data are re-computed at every iteration (Default).
+    surf_batch_size: int
+        Number of flux surfaces to compute simultaneously. Defaults to
+        computing all flux surfaces simultaneously. Decrease to reduce
+        memory required for computation.
 
     """
 
@@ -551,6 +564,7 @@ class Omnigenity(_Objective):
         "_field_data_keys",
         "_field_fixed",
         "_helicity",
+        "_surf_batch_size",
     ]
 
     _coordinates = "rtz"
@@ -577,6 +591,7 @@ class Omnigenity(_Objective):
         field_fixed=False,
         name="omnigenity",
         jac_chunk_size=None,
+        surf_batch_size=None,
     ):
         if target is None and bounds is None:
             target = 0
@@ -590,6 +605,7 @@ class Omnigenity(_Objective):
         self.eta_weight = eta_weight
         self._eq_fixed = eq_fixed
         self._field_fixed = field_fixed
+        self._surf_batch_size = surf_batch_size
         if not eq_fixed and not field_fixed:
             things = [eq, field]
         elif eq_fixed and not field_fixed:
@@ -709,6 +725,7 @@ class Omnigenity(_Objective):
             "field_transforms": field_transforms,
             "quad_weights": w,
             "helicity": self.helicity,
+            "surf_batch_size": self._surf_batch_size,
         }
 
         if self._eq_fixed:
@@ -719,6 +736,7 @@ class Omnigenity(_Objective):
                 params=self._eq.params_dict,
                 transforms=self._constants["eq_transforms"],
                 profiles=self._constants["eq_profiles"],
+                surf_batch_size=self._surf_batch_size,
             )
             self._constants["eq_data"] = eq_data
         if self._field_fixed:
@@ -730,6 +748,7 @@ class Omnigenity(_Objective):
                 transforms=self._constants["field_transforms"],
                 profiles={},
                 helicity=self._constants["helicity"],
+                surf_batch_size=self._surf_batch_size,
             )
             self._constants["field_data"] = field_data
 
@@ -789,6 +808,7 @@ class Omnigenity(_Objective):
                 params=eq_params,
                 transforms=constants["eq_transforms"],
                 profiles=constants["eq_profiles"],
+                surf_batch_size=constants["surf_batch_size"],
             )
 
         # compute field data
@@ -814,6 +834,7 @@ class Omnigenity(_Objective):
                 profiles={},
                 helicity=constants["helicity"],
                 iota=eq_data["iota"][eq_grid.unique_rho_idx],
+                surf_batch_size=constants["surf_batch_size"],
             )
             theta_B = field_data["theta_B"]
             zeta_B = field_data["zeta_B"]
@@ -840,7 +861,11 @@ class Omnigenity(_Objective):
             (field_grid.num_rho, -1)
         )
         B_mn = eq_data["|B|_mn_B"].reshape((eq_grid.num_rho, -1))
-        B_eta_alpha = vmap(_compute_B_eta_alpha)(theta_B, zeta_B, B_mn)
+        B_eta_alpha = vmap_chunked(
+            _compute_B_eta_alpha,
+            in_axes=(0, 0, 0),
+            chunk_size=constants["surf_batch_size"],
+        )(theta_B, zeta_B, B_mn)
         B_eta_alpha = B_eta_alpha.reshape(
             (field_grid.num_rho, field_grid.num_theta, field_grid.num_zeta)
         )